Extracted cleanup behaviours to another project and corretly replace LODs
