+++ Texture packs/Dolphin PC texture pack generator/Console Generator/PCTexturePackGeneratorConsole.csproj
@@ -11,10 +11,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\..\PitfallARCTool\PitfallARCTool\PitfallARCTool.csproj" />
     <ProjectReference Include="..\..\..\TexConvert\TexConvert\TexConvert.csproj" />
-    <None Update="DTETextureMap.jsonc">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </None>
+    <ProjectReference Include="..\TexturePackCleanup\TexturePackCleanup.csproj" />
   </ItemGroup>
 
 </Project>
