+++ Texture packs/Dolphin PC texture pack generator/Console Generator/bin/Release/net7.0/PitfallARCTool.runtimeconfig.dev.json
