+++ Texture packs/Dolphin PC texture pack generator/Console Generator/bin/Release/net7.0/TexturePackCleanup.deps.json
@@ -1,20 +1,20 @@
 {
   "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v5.0",
+    "name": ".NETCoreApp,Version=v7.0",
     "signature": ""
   },
   "compilationOptions": {},
   "targets": {
-    ".NETCoreApp,Version=v5.0": {
-      "PitfallARCTool/1.0.0": {
+    ".NETCoreApp,Version=v7.0": {
+      "TexturePackCleanup/1.0.0": {
         "runtime": {
-          "PitfallARCTool.dll": {}
+          "TexturePackCleanup.dll": {}
         }
       }
     }
   },
   "libraries": {
-    "PitfallARCTool/1.0.0": {
+    "TexturePackCleanup/1.0.0": {
       "type": "project",
       "serviceable": false,
       "sha512": ""
