+++ Texture packs/Dolphin PC texture pack generator/Dolphin PC texture pack generator.sln
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TexturePackCleanup", "TexturePackCleanup\TexturePackCleanup.csproj", "{9C85E81A-9E2F-4C5D-918C-E0A489B13F60}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -38,6 +40,10 @@ Global
 		{B5F9E6C1-876A-4428-8EE3-4A7EFF910776}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B5F9E6C1-876A-4428-8EE3-4A7EFF910776}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B5F9E6C1-876A-4428-8EE3-4A7EFF910776}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C85E81A-9E2F-4C5D-918C-E0A489B13F60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C85E81A-9E2F-4C5D-918C-E0A489B13F60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C85E81A-9E2F-4C5D-918C-E0A489B13F60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C85E81A-9E2F-4C5D-918C-E0A489B13F60}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
