+++ Texture packs/Dolphin PC texture pack generator/TexturePackCleanup/DTETextureMap.jsonc
