+++ Texture packs/Dolphin PC texture pack generator/TexturePackCleanup/Program.cs
@@ -1,0 +1,2 @@
+﻿// See https://aka.ms/new-console-template for more information
+Console.WriteLine("Not Yet Implemented");
